Revert "Profile/hotfix/international zip validation  (#20916)"
This reverts commit 5695fd3e96a19abf2e975d5d8e94bd35fa1b1a99.
