+++ src/applications/personalization/profile/tests/e2e/address-validation/international-zip/international-zip-is-optional.toggle-off.cypress.spec.js
@@ -1,0 +1,36 @@
+import AddressPage from '../page-objects/AddressPage';
+
+import { generateFeatureToggles } from '../../../../mocks/feature-toggles';
+
+describe('Personal and contact information', () => {
+  describe('when entering an international address', () => {
+    beforeEach(() => {
+      cy.intercept(
+        'GET',
+        '/v0/feature_toggles*',
+        generateFeatureToggles({
+          profileEnhancements: false,
+          profileDoNotRequireInternationalZipCode: false,
+        }),
+      );
+    });
+
+    it('should successfully update without zip', () => {
+      const formFields = {
+        country: 'NLD',
+        address: 'Dam 1',
+        city: 'Amsterdam',
+      };
+
+      const addressPage = new AddressPage();
+      addressPage.loadPage('international');
+      cy.injectAxeThenAxeCheck();
+      addressPage.fillAddressForm(formFields);
+      addressPage.saveForm();
+      cy.get('#root_internationalPostalCode-error-message').should(
+        'contain',
+        'Postal code is required',
+      );
+    });
+  });
+});
