--- src/applications/personalization/profile/tests/e2e/address-validation/international-zip/mocks/international-user.json
@@ -1,187 +1,0 @@
-{
-  "data": {
-    "id": "",
-    "type": "users_scaffolds",
-    "attributes": {
-      "services": ["evss-claims", "user-profile", "vet360"],
-      "account": { "accountUuid": "c049d895-ecdf-40a4-ac0f-7947a06ea0c2" },
-      "profile": {
-        "email": "vets.gov.user+36@gmail.com",
-        "firstName": "WESLEY",
-        "middleName": "WATSON",
-        "lastName": "FORD",
-        "birthDate": "1986-05-06",
-        "gender": "M",
-        "zip": "21122-6706",
-        "lastSignedIn": "2020-07-21T00:04:51.589Z",
-        "loa": { "current": 3, "highest": 3 },
-        "multifactor": true,
-        "verified": true,
-        "signIn": { "serviceName": "idme", "accountType": "N/A", "ssoe": true },
-        "authnContext": "http://idmanagement.gov/ns/assurance/loa/3"
-      },
-      "vaProfile": {
-        "status": "OK",
-        "birthDate": "19860506",
-        "familyName": "Ford",
-        "gender": "M",
-        "givenNames": ["Wesley", "Watson"],
-        "isCernerPatient": false,
-        "facilities": [{ "facilityId": "983", "isCerner": false }],
-        "vaPatient": true,
-        "mhvAccountState": "OK"
-      },
-      "veteranStatus": {
-        "status": "OK",
-        "isVeteran": true,
-        "servedInMilitary": true
-      },
-      "inProgressForms": [],
-      "prefillsAvailable": [],
-      "vet360ContactInformation": {
-        "email": {
-          "createdAt": "2020-07-30T23:38:04.000+00:00",
-          "emailAddress": "alongusername@domain.com",
-          "effectiveEndDate": null,
-          "effectiveStartDate": "2020-07-30T23:38:03.000+00:00",
-          "id": 115097,
-          "sourceDate": "2020-07-30T23:38:03.000+00:00",
-          "sourceSystemUser": null,
-          "transactionId": "604abf55-422b-4f51-b33d-9fb38b4daad1",
-          "updatedAt": "2020-07-30T23:38:04.000+00:00",
-          "vet360Id": "1273780"
-        },
-        "residentialAddress": {
-          "addressLine1": "34 Blah st",
-          "addressLine2": null,
-          "addressLine3": null,
-          "addressPou": "RESIDENCE/CHOICE",
-          "addressType": "DOMESTIC",
-          "city": "northampton",
-          "countryName": "United States",
-          "countryCodeIso2": "US",
-          "countryCodeIso3": "USA",
-          "countryCodeFips": null,
-          "countyCode": null,
-          "countyName": null,
-          "createdAt": "2020-07-09T16:18:43.000+00:00",
-          "effectiveEndDate": null,
-          "effectiveStartDate": "2020-07-09T16:18:42.000+00:00",
-          "id": 180911,
-          "internationalPostalCode": null,
-          "province": null,
-          "sourceDate": "2020-07-09T16:18:42.000+00:00",
-          "sourceSystemUser": null,
-          "stateCode": "NE",
-          "transactionId": "054c5396-9900-4bb5-ad73-4f06eb7452b5",
-          "updatedAt": "2020-07-09T16:18:43.000+00:00",
-          "validationKey": null,
-          "vet360Id": "1273780",
-          "zipCode": "05050",
-          "zipCodeSuffix": null
-        },
-        "mailingAddress": {
-          "addressLine1": "Tiburon Blvd 1708",
-          "addressLine2": null,
-          "addressLine3": null,
-          "addressPou": "CORRESPONDENCE",
-          "addressType": "INTERNATIONAL",
-          "badAddress": null,
-          "city": "Tiburon",
-          "countryCodeFips": null,
-          "countryCodeIso2": "UY",
-          "countryCodeIso3": "URY",
-          "countryName": "Uruguay",
-          "countyCode": null,
-          "countyName": null,
-          "createdAt": "2021-01-12T18:46:41.000Z",
-          "effectiveEndDate": null,
-          "effectiveStartDate": "2022-04-29T13:29:57.000Z",
-          "geocodeDate": "2022-04-29T13:29:57.000Z",
-          "geocodePrecision": 0,
-          "id": 211702,
-          "internationalPostalCode": null,
-          "latitude": 0,
-          "longitude": 0,
-          "province": null,
-          "sourceDate": "2022-04-29T13:29:57.000Z",
-          "sourceSystemUser": null,
-          "stateCode": null,
-          "transactionId": "6f3e7c8f-6ec2-4f31-b9c1-febc42a69ad9",
-          "updatedAt": "2022-04-29T13:29:58.000Z",
-          "validationKey": null,
-          "vet360Id": "1550116",
-          "zipCode": null,
-          "zipCodeSuffix": null
-        },
-        "mobilePhone": {
-          "areaCode": "555",
-          "countryCode": "1",
-          "createdAt": "2020-06-16T17:59:02.000+00:00",
-          "extension": null,
-          "effectiveEndDate": null,
-          "effectiveStartDate": "2020-07-10T15:18:18.000+00:00",
-          "id": 147570,
-          "isInternational": false,
-          "isTextable": null,
-          "isTextPermitted": false,
-          "isTty": null,
-          "isVoicemailable": null,
-          "phoneNumber": "5555559",
-          "phoneType": "MOBILE",
-          "sourceDate": "2020-07-10T15:18:18.000+00:00",
-          "sourceSystemUser": null,
-          "transactionId": "323140dd-e307-45d5-a400-76b038aa78ac",
-          "updatedAt": "2020-07-10T15:18:19.000+00:00",
-          "vet360Id": "1273780"
-        },
-        "homePhone": {
-          "areaCode": "804",
-          "countryCode": "1",
-          "createdAt": "2020-06-12T16:56:37.000+00:00",
-          "extension": "17477",
-          "effectiveEndDate": null,
-          "effectiveStartDate": "2020-07-14T19:07:45.000+00:00",
-          "id": 146766,
-          "isInternational": false,
-          "isTextable": null,
-          "isTextPermitted": null,
-          "isTty": null,
-          "isVoicemailable": null,
-          "phoneNumber": "2054422",
-          "phoneType": "HOME",
-          "sourceDate": "2020-07-14T19:07:45.000+00:00",
-          "sourceSystemUser": null,
-          "transactionId": "92c49d39-22b2-4bd6-92b4-0b7e7c63c6a9",
-          "updatedAt": "2020-07-14T19:07:46.000+00:00",
-          "vet360Id": "1273780"
-        },
-        "workPhone": {
-          "areaCode": "212",
-          "countryCode": "1",
-          "createdAt": "2020-06-25T13:30:12.000+00:00",
-          "extension": null,
-          "effectiveEndDate": null,
-          "effectiveStartDate": "2020-07-15T18:41:57.000+00:00",
-          "id": 148984,
-          "isInternational": false,
-          "isTextable": null,
-          "isTextPermitted": null,
-          "isTty": null,
-          "isVoicemailable": null,
-          "phoneNumber": "7182112",
-          "phoneType": "WORK",
-          "sourceDate": "2020-07-15T18:41:57.000+00:00",
-          "sourceSystemUser": null,
-          "transactionId": "930706ae-20f1-4576-bafb-344040f6275e",
-          "updatedAt": "2020-07-15T18:42:59.000+00:00",
-          "vet360Id": "1273780"
-        },
-        "temporaryPhone": null,
-        "faxNumber": null,
-        "textPermission": null
-      }
-    }
-  },
-  "meta": { "errors": null }
-}
