+++ src/platform/user/profile/vap-svc/actions/ui.js
@@ -26,6 +26,7 @@ export const updateFormFieldWithSchema = (
   value,
   schema = null,
   uiSchema = null,
+  options = {},
 ) => {
   const newUiSchema = updateUiSchema(uiSchema, value);
   const { data, schema: newSchema } = updateSchemaAndData(
@@ -34,6 +35,13 @@ export const updateFormFieldWithSchema = (
     value,
     true,
   );
+
+  if (
+    options?.doNotRequireInternationalPostalCode &&
+    newUiSchema?.internationalPostalCode
+  ) {
+    newUiSchema.internationalPostalCode['ui:required'] = false;
+  }
 
   return {
     type: UPDATE_PROFILE_FORM_FIELD,
